separate settings and game sceen
